requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.2
dnspython==2.7.0
Flask==3.1.1
Flask-Github==0.1.3
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-PyMongo==3.0.1
GitHub-Flask==3.2.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauth2==1.9.0.post1
oauthlib==2.1.0
pycparser==2.22
pyfiglet==1.0.2
PyJWT==2.8.0
pymongo==4.13.0
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==1.1.0
urllib3==2.4.0
Werkzeug==3.1.3
